Move comments out of project reducer
