+++ test/unit-legacy/redux/preview-test.js
@@ -54,123 +54,3 @@ tap.test('updateProjectInfo', t => {
     });
     t.end();
 });
-
-tap.test('setComments', t => {
-    // Initial value
-    t.deepEqual(initialState.comments, []);
-
-    state = reducer(initialState, Preview.setComments([{id: 1}, {id: 2}]));
-    state = reducer(state, Preview.setComments([{id: 3}, {id: 4}]));
-    t.deepEqual(state.comments, [{id: 1}, {id: 2}, {id: 3}, {id: 4}]);
-
-    t.end();
-});
-
-const commentState = {
-    comments: [
-        {id: 'id1', visibility: 'visible'},
-        {id: 'id2', visibility: 'visible'},
-        {id: 'id3', visibility: 'visible'}
-    ],
-    replies: {
-        id1: [
-            {id: 'id4', visibility: 'visible'},
-            {id: 'id5', visibility: 'visible'}
-        ]
-    }
-};
-
-tap.test('setComments, discards duplicates', t => {
-    state = reducer(commentState, Preview.setComments([{id: 'id1'}]));
-    // Does not increase the number of comments, still 3
-    t.equal(state.comments.length, 3);
-    t.end();
-});
-
-tap.test('setCommentDeleted, top level comment', t => {
-    state = reducer(commentState, Preview.setCommentDeleted('id2'));
-    t.equal(state.comments[1].visibility, 'deleted');
-    t.end();
-});
-
-tap.test('setCommentDeleted, reply comment', t => {
-    state = reducer(commentState, Preview.setCommentDeleted('id4', 'id1'));
-    t.equal(state.replies.id1[0].visibility, 'deleted');
-    t.end();
-});
-
-tap.test('setRepliesDeleted/Restored', t => {
-    state = reducer(commentState, Preview.setRepliesDeleted('id1'));
-    t.equal(state.replies.id1[0].visibility, 'deleted');
-    t.equal(state.replies.id1[1].visibility, 'deleted');
-
-    state = reducer(state, Preview.setRepliesRestored('id1'));
-    t.equal(state.replies.id1[0].visibility, 'visible');
-    t.equal(state.replies.id1[1].visibility, 'visible');
-    t.end();
-});
-
-tap.test('setCommentReported, top level comment', t => {
-    state = reducer(commentState, Preview.setCommentReported('id2'));
-    t.equal(state.comments[1].visibility, 'reported');
-    t.end();
-});
-
-tap.test('setCommentReported, reply comment', t => {
-    state = reducer(commentState, Preview.setCommentReported('id4', 'id1'));
-    t.equal(state.replies.id1[0].visibility, 'reported');
-    t.end();
-});
-
-tap.test('addNewComment, top level comment', t => {
-    state = reducer(commentState, Preview.addNewComment({id: 'new comment'}));
-    // Adds comment to beginning of list
-    t.equal(state.comments[0].id, 'new comment');
-    t.end();
-});
-
-tap.test('addNewComment, reply comment', t => {
-    state = reducer(commentState, Preview.addNewComment({id: 'new comment'}, 'id1'));
-    // Adds replies to the end of the replies list
-    t.equal(state.replies.id1[2].id, 'new comment');
-    t.end();
-});
-
-tap.test('setReplies', t => {
-    // setReplies should append new replies
-    state = reducer(commentState, Preview.setReplies({
-        id1: {id: 'id6'}
-    }));
-    t.equal(state.replies.id1[2].id, 'id6');
-    t.equal(state.comments[0].moreRepliesToLoad, false);
-
-    // setReplies should ignore duplicates, do the same as above again
-    t.equal(state.replies.id1.length, 3);
-    state = reducer(state, Preview.setReplies({id1: {id: 'id6'}}));
-    t.equal(state.replies.id1.length, 3);
-
-    // setReplies can add replies to a comment that didn't have any
-    state = reducer(state, Preview.setReplies({
-        id2: {id: 'id7'}
-    }));
-    t.equal(state.replies.id1.length, 3);
-    t.equal(state.replies.id2.length, 1);
-    t.equal(state.replies.id2[0].id, 'id7');
-    t.equal(state.comments[0].moreRepliesToLoad, false);
-    t.equal(state.comments[1].moreRepliesToLoad, false);
-
-    // Getting 20 (COMMENT_LIMIT) replies sets moreRepliesToLoad to true
-    state = reducer(state, Preview.setReplies({
-        id3: (new Array(20)).map((_, i) => ({id: `id${i + 1}`}))
-    }));
-    t.equal(state.comments[0].moreRepliesToLoad, false);
-    t.equal(state.comments[1].moreRepliesToLoad, false);
-    t.equal(state.comments[2].moreRepliesToLoad, true);
-
-    // Getting one more reply sets moreRepliesToLoad back to false
-    state = reducer(state, Preview.setReplies({
-        id3: {id: 'id21'}
-    }));
-    t.equal(state.comments[2].moreRepliesToLoad, false);
-    t.end();
-});
